chat-server/pyproject.toml
[tool.poetry]
name = "chat-server"
version = "0.1.0"
description = ""
authors = ["MinJuTur <jmj4701jmj@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.127.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
SQLAlchemy = ">=1.4,<2.0"
psycopg2-binary = "^2.9.11"
valkey-glide = "^2.2.3"
python-dotenv = "^1.2.1"
websockets = "^15.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
